// src/types.ts
export type SeedState = 'open' | 'closed';

// Method selection
export type DivinationMethod = 'opele' | 'opon' | 'merindilogun';

export interface OpeleState {
  rightLeg: [SeedState, SeedState, SeedState, SeedState];
  leftLeg: [SeedState, SeedState, SeedState, SeedState];
}

// State for 16 Cowries
export type CowrieState = SeedState[]; // Array of 16 states

export interface OduInfo {
  name: string;
  isMeji: boolean; 
  binaryRepresentation: string; 
}